+++ .vs/GameServer/xs/UserPrefs.xml
@@ -1,21 +1,15 @@
 ﻿<Properties StartupConfiguration="{F9F77F06-66D7-EF2A-06E6-48CD1F3E7407}|">
-  <MonoDevelop.Ide.Workbench ActiveDocument="Assets/src/Game/GameController.cs">
+  <MonoDevelop.Ide.Workbench ActiveDocument="Assets/src/Game/UserController.cs">
     <Files>
       <File FileName="Assets/src/Constant.cs" Line="22" Column="1" />
-      <File FileName="Assets/src/Game/GameController.cs" Line="81" Column="94" />
+      <File FileName="Assets/src/Game/GameController.cs" Line="111" Column="1" />
       <File FileName="Assets/src/Game/TCP_ServerController.cs" />
-      <File FileName="Assets/src/Game/UDP_ServerController.cs" />
-      <File FileName="Assets/src/Game/UserAnimation.cs" />
-      <File FileName="Assets/src/Game/UserController.cs" Line="294" Column="9" />
+      <File FileName="Assets/src/Game/UDP_ServerController.cs" Line="71" Column="1" />
+      <File FileName="Assets/src/Game/UserAnimation.cs" Line="8" Column="1" />
+      <File FileName="Assets/src/Game/UserController.cs" Line="400" Column="28" />
       <File FileName="Assets/src/Game/HeaderClass.cs" />
       <File FileName="Assets/src/Game/Bom/Grenade.cs" />
-      <File FileName="Assets/src/Game/Bom/Explosion.cs" />
-      <File FileName="Assets/src/Library/TimeMeasurment.cs" />
-      <File FileName="Assets/src/Library/OpenSocket/TCP_Server.cs" />
-      <File FileName="Assets/src/Library/OpenSocket/UDP_Server.cs" />
-      <File FileName="Assets/src/Library/Convert.cs" />
-      <File FileName="Assets/src/Library/BaseWeapon.cs" />
-      <File FileName="Assets/src/Library/StateMachine/StateMachine.cs" />
+      <File FileName="Assets/src/Game/Communication/UserBodyData.cs" Line="1" Column="1" />
     </Files>
     <Pads>
       <Pad Id="ProjectPad">
@@ -25,7 +19,8 @@
               <Node name="src" expanded="True">
                 <Node name="Game" expanded="True">
                   <Node name="Bom" expanded="True" />
-                  <Node name="GameController.cs" selected="True" />
+                  <Node name="Communication" expanded="True" />
+                  <Node name="UserController.cs" selected="True" />
                 </Node>
                 <Node name="Library" expanded="True">
                   <Node name="Animator" expanded="True" />
@@ -41,7 +36,7 @@
   </MonoDevelop.Ide.Workbench>
   <MonoDevelop.Ide.DebuggingService.PinnedWatches />
   <MonoDevelop.Ide.Workspace ActiveConfiguration="Debug" />
-  <MonoDevelop.Ide.ItemProperties.Assembly-CSharp PreferredExecutionTarget="1724fc89-fe8b-43e8-bb48-f2227bc454c1" />
+  <MonoDevelop.Ide.ItemProperties.Assembly-CSharp PreferredExecutionTarget="252c341b-3264-4d13-a498-84e4464da2ed" />
   <MonoDevelop.Ide.DebuggingService.Breakpoints>
     <BreakpointStore />
   </MonoDevelop.Ide.DebuggingService.Breakpoints>
