+++ Assets/src/Game/Communication.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1aa36827867404bb98af641803f6ecbc
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
