+++ Assets/src/Game/Communication/HeaderClass.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bf73bcb15115f4fa09aa5ff8a5f9d44b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
