+++ Assets/src/Game/Communication/UserBodyData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3532dd3d32a5148a284e031e57448901
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
