--- Assets/src/Game/TCP_ServerController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cbe70584f59e14fecaa5cd3e1d964b3c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
