--- Assets/src/Game/UDP_ServerController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ffe99439dd8af4b158a81021bb6e8645
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
