+++ Assets/src/Game/UserController.cs
@@ -8,6 +8,9 @@ using System.Threading.Tasks;
 
 public class UserController : MonoBehaviour
 {
+    UserBodyData userData = new UserBodyData();
+
+
     public string userId;
     public string IPaddr;
     public uint sequence=0;
@@ -154,24 +157,28 @@ public class UserController : MonoBehaviour
         return returnData;
     }
 
-    
+
     public byte[] GetStatus()
     {
         List<byte> returnData = new List<byte>();
         GameHeader header = new GameHeader();
         header.CreateNewData(GameHeader.ID.GAME, this.name, (byte)GameHeader.GameCode.BASICDATA);
-
-        byte[] positionData = Convert.GetByteVector3(this.transform.position);
-        byte[] rotationData = Convert.GetByteVector3(this.transform.localEulerAngles);
-        int currentKey = 0;
-        if (userAnimation) currentKey = (int)userAnimation.animationState.currentKey;
-
+        userData.SetData(this.transform.position, this.transform.localEulerAngles, (int)userAnimation.animationState.currentKey, hp);
         returnData.AddRange(header.GetHeader());
-        returnData.AddRange(positionData);                              //16byte
-        returnData.AddRange(rotationData);                              //32byte
-        returnData.AddRange(BitConverter.GetBytes(currentKey));         //34byte
-        returnData.AddRange(BitConverter.GetBytes(hp));                 //38byte
-        if (weapon != null) returnData.AddRange(weapon.GetStatus());    //50byte
+        returnData.AddRange(userData.GetData());
+        if (weapon != null) returnData.AddRange(weapon.GetStatus());
+        return returnData.ToArray();
+    }
+
+    public byte[] GetStatusComplete()
+    {
+        List<byte> returnData = new List<byte>();
+        GameHeader header = new GameHeader();
+        header.CreateNewData(GameHeader.ID.GAME, this.name, (byte)GameHeader.GameCode.BASICDATA);
+        userData.SetData(this.transform.position, this.transform.localEulerAngles, (int)userAnimation.animationState.currentKey, hp);
+        returnData.AddRange(header.GetHeader());
+        returnData.AddRange(userData.GetCompleteData());
+        if (weapon != null) returnData.AddRange(weapon.GetStatus());
         return returnData.ToArray();
 
     }
@@ -279,115 +286,3 @@ public class UserController : MonoBehaviour
     }
 }
 
-public class UserBodyData
-{
-    public enum USERDATAFLG : byte
-    {
-        POSITION_X= 0x001,
-        POSITION_Y= 0x002,
-        POSITION_Z= 0x004,
-        ROTATION_X= 0x008,
-        ROTATION_Y = 0x010,
-        ROTATION_Z = 0x020,        
-    }
-
-    
-    public Vector3 position = Vector3.zero;
-    public Vector3 rotetion = Vector3.zero;
-    public int animationKey =0;
-    public int hp = 0;
-
-    private USERDATAFLG sendDataFlg = 0x000;
-
-    void SetData(Vector3 _position, Vector3 _rotetion, int _currentKey, int _hp)
-    {
-        //position
-        {
-            bool flg = false;
-            if (Math.Abs(position.x-_position.x)>0.001)
-            {
-                sendDataFlg |= USERDATAFLG.POSITION_X;
-                flg = true;
-            }
-            if (Math.Abs(position.y - _position.y) > 0.001)
-            {
-                sendDataFlg |= USERDATAFLG.POSITION_Y;
-                flg = true;
-            }
-            if (Math.Abs(position.y - _position.y) > 0.001)
-            {
-                sendDataFlg |= USERDATAFLG.POSITION_Z;
-                flg = true;
-            }
-            if (flg)position = _position;
-
-        }
-
-        //rotation
-        {
-            bool flg = false;
-            if (Math.Abs(rotetion.x - _rotetion.x) > 0.001)
-            {
-                sendDataFlg |= USERDATAFLG.ROTATION_X;
-                flg = true;
-            }
-            if (Math.Abs(rotetion.y - _rotetion.y) > 0.001)
-            {
-                sendDataFlg |= USERDATAFLG.ROTATION_Y;
-                flg = true;
-            }
-            if (Math.Abs(rotetion.y - _rotetion.y) > 0.001)
-            {
-                sendDataFlg |= USERDATAFLG.ROTATION_Z;
-                flg = true;
-            }
-            if (flg)rotetion = _rotetion;
-
-        }
-
-        animationKey = _currentKey;
-        hp = _hp;
-    }
-
-    byte[] GetData(UserBodyData _oldData = null)
-    {
-
-        List<byte> returnData = new List<byte>();
-
-        //送信データ作成
-        returnData.Add((byte)sendDataFlg);
-        if (sendDataFlg.HasFlag(USERDATAFLG.POSITION_X)) returnData.AddRange(BitConverter.GetBytes(position.x));
-        if (sendDataFlg.HasFlag(USERDATAFLG.POSITION_Y)) returnData.AddRange(BitConverter.GetBytes(position.y));
-        if (sendDataFlg.HasFlag(USERDATAFLG.POSITION_Z)) returnData.AddRange(BitConverter.GetBytes(position.z));
-        if (sendDataFlg.HasFlag(USERDATAFLG.ROTATION_X)) returnData.AddRange(BitConverter.GetBytes(rotetion.x));
-        if (sendDataFlg.HasFlag(USERDATAFLG.ROTATION_Y)) returnData.AddRange(BitConverter.GetBytes(rotetion.y));
-        if (sendDataFlg.HasFlag(USERDATAFLG.ROTATION_Z)) returnData.AddRange(BitConverter.GetBytes(rotetion.z));
-
-        returnData.AddRange(BitConverter.GetBytes(animationKey));
-        returnData.AddRange(BitConverter.GetBytes(hp));
-        sendDataFlg = 0;
-
-
-        return returnData.ToArray();
-    }
-
-    void Deserialize(byte[] _data,int _index)
-    {
-        int nowIndex=_index;
-        USERDATAFLG nowFlg = (USERDATAFLG)_data[_index++];
-        if (nowFlg.HasFlag(USERDATAFLG.POSITION_X))position.x = GetFloat(_data,ref nowIndex);
-        if (nowFlg.HasFlag(USERDATAFLG.POSITION_Y)) position.y = GetFloat(_data, ref nowIndex);
-        if (nowFlg.HasFlag(USERDATAFLG.POSITION_Z)) position.z = GetFloat(_data, ref nowIndex);
-        if (nowFlg.HasFlag(USERDATAFLG.ROTATION_X)) rotetion.x = GetFloat(_data, ref nowIndex);
-        if (nowFlg.HasFlag(USERDATAFLG.ROTATION_Y)) rotetion.y = GetFloat(_data, ref nowIndex);
-        if (nowFlg.HasFlag(USERDATAFLG.ROTATION_Z)) rotetion.z = GetFloat(_data, ref nowIndex);
-        animationKey=BitConverter.ToInt32(_data, nowIndex);
-        nowIndex += sizeof(int);
-        hp = BitConverter.ToInt32(_data, nowIndex);
-    }
-    private float GetFloat(byte[] _data,ref int _startIndex)
-    {
-        _startIndex += 4;
-        return BitConverter.ToSingle(_data, _startIndex-4);
-    }
-}
